feat: add a method for capturing Hyprland clients
feat: improve project's structure
