+++ glace/glace-private.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <fcntl.h>
+#include <sys/mman.h>
+
+#include "hyprland-toplevel-export-v1.h"
+#include "glace-client-private.h"
+#include "glace-manager-private.h"
+#include "wlr-foreign-toplevel-management-unstable-v1.h"
