+++ glace/glace.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "glace-client.h"
+#include "glace-manager.h"
