--- libglace/libglace-private.h
@@ -1,2 +1,0 @@
-#include "libglace-client-private.h"
-#include "libglace-manager-private.h"
