--- libglace/libglace.c
@@ -1,2 +1,0 @@
-#include "libglace-manager.h"
-#include "libglace-client.h"
